remove lib/auth/test package
